+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from "./harmaja"
 export * from "./listview"
 export * from "./jsxfactory"
-export * from "./observables"
+export * from "./signal"
