+++ src/signal/lens.ts
@@ -1,0 +1,4 @@
+export interface Lens<A, B> {
+    get(root: A): B
+    set(root: A, newValue: B): A
+}
