+++ src/signal/observer.ts
@@ -1,0 +1,2 @@
+export type Observer<T> = (newValue: T) => void
+export type Unsubscribe = () => void
